Выполненное ДЗ_4
Перевёл приложение на модули, добавил интерфейсов, переделал чтение файла students.
