+++ learning-center-impl/src/main/java/space/irsi7/Main.java
