+++ learning-center-impl/src/main/java/space/irsi7/controllers/MultipleChoiceMenu.java
