+++ learning-center-impl/src/main/java/space/irsi7/dao/YamlDaoImpl.java
@@ -1,0 +1,65 @@
+package space.irsi7.dao;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
+import org.yaml.snakeyaml.Yaml;
+import space.irsi7.interfaces.YamlDAO;
+import space.irsi7.models.Course;
+import space.irsi7.models.Student;
+import space.irsi7.models.Theme;
+
+import java.io.*;
+import java.util.*;
+
+public class YamlDaoImpl implements YamlDAO {
+
+    ObjectMapper mapper = new YAMLMapper().enable(SerializationFeature.INDENT_OUTPUT);
+    Yaml yaml = new Yaml();
+
+    public void writeYAML(List<Object> arrayList, String path) throws IOException {
+        FileWriter writer = new FileWriter(path);
+        yaml.dump(arrayList, writer);
+    }
+
+    public Map<Integer, Student> readYamlStudents(String path) throws IOException {
+
+        Map<Integer, Student> answer = new HashMap<>();
+        ArrayList<?> studentsFile = mapper.readValue(new File(path), new TypeReference<>() {} );
+        studentsFile.forEach(sc -> {
+            if (sc instanceof Map<?, ?> someClass) {
+                Student student = new Student(someClass);
+                answer.put(student.getId(), student);
+            }
+        });
+        return answer;
+    }
+
+
+//    public Map<Integer, Student> getStudentsMap(String path) throws IOException {
+//        return readYamlStudentsArray(path).stream().collect(Collectors.toMap(Student::getId, s -> s));
+//    }
+
+    public void readYamlConfig(String path,
+                               Map<Integer, Course> courseMap,
+                               Map<Integer, Theme> themeMap) throws IOException {
+        Map<String, List<?>> configFile = mapper.readValue(new File(path), new TypeReference<>() {} );
+
+        // Заполнение мапы courseMap
+        configFile.get("courses").forEach(sc -> {
+            if (sc instanceof Map<?, ?> someClass) {
+                Course course = new Course(someClass);
+                courseMap.put(course.id, course);
+            }
+        });
+
+        // Заполнение мапы themeMap
+        configFile.get("themes").forEach(sc -> {
+            if (sc instanceof Map<?, ?> someClass) {
+                Theme theme = new Theme(someClass);
+                themeMap.put(theme.id, theme);
+            }
+        });
+    }
+}
