+++ learning-center-impl/src/main/java/space/irsi7/enums/LogsEnum.java
