+++ learning-center-impl/src/main/java/space/irsi7/enums/MenuEnum.java
