+++ learning-center-impl/src/main/resources/data/properties.yaml
