+++ learning-center-impl/src/main/resources/data/students.yaml
@@ -1,0 +1,42 @@
+- !!space.irsi7.models.Student
+  courseId: 1
+  gpa: 5
+  id: 1
+  marks: [4, 5, 5]
+  name: Roman Ivanov
+- !!space.irsi7.models.Student
+  courseId: 2
+  gpa: 22
+  id: 2
+  marks: [4, 5, 5, 5, 75, 2, 60]
+  name: Sasha Zaycev
+- !!space.irsi7.models.Student
+  courseId: 2
+  gpa: 83
+  id: 3
+  marks: [80, 95, 75]
+  name: Vlad Neglad
+- !!space.irsi7.models.Student
+  courseId: 1
+  gpa: 40
+  id: 4
+  marks: [40]
+  name: Stepan Smith
+- !!space.irsi7.models.Student
+  courseId: 1
+  gpa: 0
+  id: 5
+  marks: [10]
+  name: John Smit
+- !!space.irsi7.models.Student
+  courseId: 2
+  gpa: 95
+  id: 6
+  marks: [100, 90]
+  name: John Smith
+- !!space.irsi7.models.Student
+  courseId: 1
+  gpa: 0
+  id: 7
+  marks: []
+  name: Emily Smith
