+++ learning-center-model/src/main/java/space/irsi7/exceptions/IllegalInitialDataException.java
