+++ learning-center-model/src/main/java/space/irsi7/models/Course.java
@@ -1,0 +1,18 @@
+package space.irsi7.models;
+
+import java.util.*;
+
+public class Course extends Readable {
+
+    public int id;
+    public ArrayList<Integer> themeIds;
+
+    public Course(int id, ArrayList<Integer> themeIds){
+        this.id = id;
+        this.themeIds = themeIds;
+    }
+
+    public Course(Map<?, ?> course){
+        super(course);
+    }
+}
