+++ learning-center-model/src/main/resources/archetype-resources/src/main/java/App.java
@@ -1,0 +1,13 @@
+package $space.irsi7;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
