+++ logs/app-log.log
@@ -1,0 +1,16 @@
+16:13:18.409 [main] INFO  space.irsi7.Main - Приложение запущено
+16:13:18.584 [main] INFO  s.i.repository.StudentRepositoryImpl - Данные о студентах успешно считаны из students.yaml
+16:13:18.592 [main] INFO  s.irsi7.services.StudentServiceImpl - Данные о темах и курсах успешно считаны из config.yaml
+16:13:49.249 [main] INFO  space.irsi7.Main - Приложение запущено
+16:13:49.418 [main] INFO  s.i.repository.StudentRepositoryImpl - Данные о студентах успешно считаны из students.yaml
+16:13:49.425 [main] INFO  s.irsi7.services.StudentServiceImpl - Данные о темах и курсах успешно считаны из config.yaml
+16:14:07.353 [main] INFO  space.irsi7.Main - Приложение завершило работу
+16:15:01.946 [main] INFO  space.irsi7.Main - Приложение запущено
+16:15:02.119 [main] INFO  s.i.repository.StudentRepositoryImpl - Данные о студентах успешно считаны из students.yaml
+16:15:02.127 [main] INFO  s.irsi7.services.StudentServiceImpl - Данные о темах и курсах успешно считаны из config.yaml
+16:15:28.443 [main] INFO  space.irsi7.Main - РџСЂРёР»РѕР¶РµРЅРёРµ Р·Р°РїСѓС‰РµРЅРѕ
+16:15:28.821 [main] INFO  s.i.repository.StudentRepositoryImpl - Р”Р°РЅРЅС‹Рµ Рѕ СЃС‚СѓРґРµРЅС‚Р°С… СѓСЃРїРµС€РЅРѕ СЃС‡РёС‚Р°РЅС‹ РёР· students.yaml
+16:15:28.834 [main] INFO  s.irsi7.services.StudentServiceImpl - Р”Р°РЅРЅС‹Рµ Рѕ С‚РµРјР°С… Рё РєСѓСЂСЃР°С… СѓСЃРїРµС€РЅРѕ СЃС‡РёС‚Р°РЅС‹ РёР· config.yaml
+16:16:23.158 [main] INFO  space.irsi7.Main - РџСЂРёР»РѕР¶РµРЅРёРµ Р·Р°РїСѓС‰РµРЅРѕ
+16:16:35.782 [main] INFO  s.i.repository.StudentRepositoryImpl - Р”Р°РЅРЅС‹Рµ Рѕ СЃС‚СѓРґРµРЅС‚Р°С… СѓСЃРїРµС€РЅРѕ СЃС‡РёС‚Р°РЅС‹ РёР· students.yaml
+16:16:42.244 [main] INFO  s.irsi7.services.StudentServiceImpl - Р”Р°РЅРЅС‹Рµ Рѕ С‚РµРјР°С… Рё РєСѓСЂСЃР°С… СѓСЃРїРµС€РЅРѕ СЃС‡РёС‚Р°РЅС‹ РёР· config.yaml
