+++ pom.xml
@@ -7,12 +7,19 @@
     <groupId>space.irsi7</groupId>
     <artifactId>StudentsEducationSystem</artifactId>
     <version>1.0-SNAPSHOT</version>
+    <modules>
+        <module>learning-center-api</module>
+        <module>learning-center-model</module>
+        <module>learning-center-impl</module>
+    </modules>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
         <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
+
+    <packaging>pom</packaging>
 
     <dependencies>
 
@@ -49,4 +56,5 @@
             <version>1.2.9</version>
         </dependency>
     </dependencies>
+
 </project>
