--- src/main/java/space/irsi7/enums/PathsEnum.java
@@ -1,16 +1,0 @@
-package space.irsi7.enums;
-
-public enum PathsEnum {
-    STUDENTS("src/main/resources/data/students.yaml"),
-    CONFIG("src/main/resources/data/properties.yaml");
-
-    public String getPath() {
-        return path;
-    }
-
-    private final String path;
-
-    PathsEnum(String path) {
-        this.path = path;
-    }
-}
