--- src/main/resources/data/students.yaml
@@ -1,36 +1,0 @@
-- !!space.irsi7.models.Student
-  courseId: 1
-  gpa: 5
-  id: 1
-  marks: [4, 5, 5]
-  name: Roman Ivanov
-- !!space.irsi7.models.Student
-  courseId: 2
-  gpa: 16
-  id: 2
-  marks: [4, 5, 5, 5, 75, 2]
-  name: Sasha Zaycev
-- !!space.irsi7.models.Student
-  courseId: 2
-  gpa: 83
-  id: 3
-  marks: [80, 95, 75]
-  name: Vlad Neglad
-- !!space.irsi7.models.Student
-  courseId: 1
-  gpa: 40
-  id: 4
-  marks: [40]
-  name: Stepan Smit
-- !!space.irsi7.models.Student
-  courseId: 1
-  gpa: 0
-  id: 5
-  marks: []
-  name: John Smit
-- !!space.irsi7.models.Student
-  courseId: 2
-  gpa: 95
-  id: 6
-  marks: [100, 90]
-  name: Марк Заславский
